+++ frontend/src/components/SpotsIndex.css
@@ -22,4 +22,19 @@
 .notLink {
     text-decoration: none;
     color: black;
+    
+}
+
+.notLinkPrice {
+    font-weight: bold;
+}
+
+.toolTipText {
+    visibility: hidden;
+    color: black;
+}
+
+.toolTipText:hover {
+    visibility: visible;
+    opacity: 1;
 }
