+++ frontend/src/components/SpotsShow.js
@@ -10,24 +10,14 @@ import {
 } from "../store/Reviews";
 import OpenModalButton from "./OpenModalButton";
 import { useModal } from "../context/Modal";
+import CreateReview from "./CreateReview";
 
 const SpotsShow = () => {
   const { spotId } = useParams();
   const dispatch = useDispatch();
   const { closeModal } = useModal();
 
-  const [rating, setRating] = useState(0);
-  console.log("🚀 ~ file: SpotsShow.js:16 ~ SpotsShow ~ rating:", rating);
-  const [reviewText, setReviewText] = useState("");
-  console.log(
-    "🚀 ~ file: SpotsShow.js:17 ~ SpotsShow ~ reviewText:",
-    reviewText
-  );
-  const [hoverRating, setHoverRating] = useState(0);
-  console.log(
-    "🚀 ~ file: SpotsShow.js:18 ~ SpotsShow ~ hoverRating:",
-    hoverRating
-  );
+  
   const [errors, setErrors] = useState({});
 
   useEffect(() => {
@@ -37,7 +27,7 @@ const SpotsShow = () => {
 
   const spot = useSelector((state) => state.spots[spotId]);
   const data = { ...spot };
-  console.log("🚀 ~ file: SpotsShow.js:18 ~ SpotsShow ~ data:", data)
+//   console.log("🚀 ~ file: SpotsShow.js:18 ~ SpotsShow ~ data:", data)
 
   const review1 = useSelector((state) => state.reviews);
   const reviewArray = Object.values(review1);
@@ -46,30 +36,7 @@ const SpotsShow = () => {
   const session = useSelector((state) => state.session.user);
   // console.log("🚀 ~ file: SpotsShow.js:39 ~ SpotsShow ~ session:", session)
 
-  async function onSubmit(e) {
-    e.preventDefault();
-
-    const reviewForm = {
-      review: reviewText,
-      stars: rating,
-    };
-    console.log(
-      "🚀 ~ file: SpotsShow.js:38 ~ onSubmit ~ reviewForm :",
-      reviewForm
-    );
-    const res = dispatch(thunkCreateReview(reviewForm, spotId))
-      .then(closeModal())
-      .catch(errors);
-    if (!res.errors) {
-      closeModal();
-    } else {
-      setErrors(res.errors);
-      console.log(
-        "🚀 ~ file: SpotsShow.js:56 ~ onSubmit ~ res.errors:",
-        res.errors
-      );
-    }
-  }
+ 
 
   return (
     <div className="spotsShowContainer">
@@ -129,67 +96,7 @@ const SpotsShow = () => {
         hidden={data.ownerId === session.id}
         buttonText="Post Your Review"
         modalComponent={
-          <div className="postReviewModal">
-            <h1>How was your stay?</h1>
-            <p>{errors.review}</p>
-            <p>{errors.stars}</p>
-            <form onSubmit={onSubmit}>
-              <label>
-                <input
-                  className="postReviewModalTextbox"
-                  type="text"
-                  value={reviewText}
-                  placeholder="Just a quick review."
-                  onChange={(e) => setReviewText(e.target.value)}
-                ></input>
-              </label>
-              <div className="rating">
-                <div
-                  onMouseEnter={() => setHoverRating(1)}
-                  onMouseLeave={() => setHoverRating(0)}
-                  onClick={() => setRating(1)}
-                  className={hoverRating >= 1 ? "full" : "empty"}
-                >
-                  <i className="fa fa-star"></i>
-                </div>
-                <div
-                  onMouseEnter={() => setHoverRating(2)}
-                  onMouseLeave={() => setHoverRating(0)}
-                  onClick={() => setRating(2)}
-                  className={hoverRating >= 2 ? "full" : "empty"}
-                >
-                  <i className="fa fa-star"></i>
-                </div>
-                <div
-                  onMouseEnter={() => setHoverRating(3)}
-                  onMouseLeave={() => setHoverRating(0)}
-                  onClick={() => setRating(3)}
-                  className={hoverRating >= 3 ? "full" : "empty"}
-                >
-                  <i className="fa fa-star"></i>
-                </div>
-                <div
-                  onMouseEnter={() => setHoverRating(4)}
-                  onMouseLeave={() => setHoverRating(0)}
-                  onClick={() => setRating(4)}
-                  className={hoverRating >= 3 ? "full" : "empty"}
-                >
-                  <i className="fa fa-star"></i>
-                </div>
-                <div
-                  onMouseEnter={() => setHoverRating(5)}
-                  onMouseLeave={() => setHoverRating(0)}
-                  onClick={() => setRating(5)}
-                  className={hoverRating >= 5 ? "full" : "empty"}
-                >
-                  <i className="fa fa-star"></i>
-                </div>
-              </div>
-              <button className="postReviewModalSubmit" type="submit">
-                Submit Your Review
-              </button>
-            </form>
-          </div>
+          <CreateReview spotId={spotId}/>
         }
       />
     </button>
